render slog fragment
