+++ packages/radix-vue/src/shared/renderSlotFragments.ts
@@ -1,0 +1,11 @@
+import { Fragment, type VNode } from "vue";
+
+export function renderSlotFragments(children?: VNode[]): VNode[] {
+  if (!children) return [];
+  return children.flatMap((child) => {
+    if (child.type === Fragment) {
+      return renderSlotFragments(child.children as VNode[]);
+    }
+    return [child];
+  });
+}
